Refactor name of sample to Orbit Camera Controller
