+++ src/main/java/com/esri/samples/scene/orbit_camera_controller/ControlTheCamera.png
