+++ src/main/resources/fxml/orbit_camera_controller.fxml
@@ -23,7 +23,7 @@
 <?import com.esri.arcgisruntime.mapping.view.SceneView?>
 
 <?import javafx.geometry.Insets?>
-<StackPane fx:controller="com.esri.samples.scene.control_the_camera.ControlTheCameraController"
+<StackPane fx:controller="com.esri.samples.scene.orbit_camera_controller.OrbitCameraControllerController"
            xmlns:fx="http://javafx.com/fxml"
            stylesheets="/css/style.css">
